Optimize SPO2 algorithm

+++ SW/03-common/Platform_TYPES.h
@@ -36,8 +36,8 @@
 /******** Data Types **********/
 typedef unsigned char       boolean;
 typedef unsigned char       uint8 ;
-typedef unsigned short int  uint16;
-typedef unsigned long  int  uint32;
+typedef unsigned int  			uint16;
+typedef unsigned long int  	uint32;
 typedef unsigned long long  uint64;
 
 typedef signed char         sint8 ;
